pyproject.toml
[project]
name = "graphbot"
dynamic = ["version"]
description = "LangGraph-based AI assistant framework"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "GraphBot Contributors"}]
keywords = ["langgraph", "agent", "llm", "fastapi", "ai-assistant", "chatbot", "sqlite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # LangGraph & LangChain
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # LLM Providers
    "litellm>=1.0.0",
    "openrouter>=0.7.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",

    # Background
    "apscheduler>=3.10.0",
    "croniter>=1.3.0",

    # Web search
    "duckduckgo-search>=8.0.0",

    # Logging
    "loguru>=0.7.0",

    # Auth
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.urls]
Repository = "https://github.com/omrylcn/graphbot"
Changelog = "https://github.com/omrylcn/graphbot/blob/main/CHANGELOG.md"

[project.scripts]
gbot = "gbot_cli.commands:app"
graphbot = "gbot_cli.commands:app"

[project.optional-dependencies]
channels = [
    "python-telegram-bot>=21.0",
    "websockets>=12.0",
]
rag = [
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "graphbot/__version__.py"

[tool.hatch.build.targets.wheel]
packages = ["graphbot", "gbot_cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: Integration tests (require running server on localhost:8000)",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.5.4",
]
